This is going to my initial commit.
What is working and what I am working on

Working.
AR Camera running both a food not food model using vision. Then if the vision model detects a food running a food model over that.

Not Working.
I have recently changed a lot of the UI to make it a look a little better. So yeah there are some buttons there that just don't do anything.

This is the first to a lot of commits to come.